Remove duplicated collideable class
